+++ woodwork/column_accessor.py
@@ -20,7 +20,11 @@ from woodwork.indexers import _iLocIndexer, _locIndexer
 from woodwork.logical_types import _NULLABLE_PHYSICAL_TYPES, LatLong, Ordinal
 from woodwork.statistics_utils import _get_box_plot_info_for_column
 from woodwork.table_schema import TableSchema
-from woodwork.utils import _get_column_logical_type, import_or_none
+from woodwork.utils import (
+    _check_data_type_equality,
+    _get_column_logical_type,
+    import_or_none,
+)
 
 dd = import_or_none("dask.dataframe")
 ps = import_or_none("pyspark.pandas")
@@ -103,7 +107,9 @@ class WoodworkColumnAccessor:
                     logical_type.validate(self._series)
                 else:
                     valid_dtype = logical_type._get_valid_dtype(type(self._series))
-                    if valid_dtype != str(self._series.dtype):
+                    if not _check_data_type_equality(
+                        valid_dtype, str(self._series.dtype)
+                    ):
                         raise TypeValidationError(
                             f"Cannot initialize Woodwork. Series dtype '{self._series.dtype}' is "
                             f"incompatible with {logical_type} dtype. Try converting series "
@@ -280,7 +286,7 @@ class WoodworkColumnAccessor:
                     valid_dtype = self._schema.logical_type._get_valid_dtype(
                         type(result)
                     )
-                    if str(result.dtype) == valid_dtype:
+                    if _check_data_type_equality(str(result.dtype), valid_dtype):
                         result.ww.init(schema=self.schema, validate=False)
                     else:
                         invalid_schema_message = (
@@ -446,7 +452,7 @@ def _validate_schema(schema, series):
         raise TypeError("Provided schema must be a Woodwork.ColumnSchema object.")
 
     valid_dtype = schema.logical_type._get_valid_dtype(type(series))
-    if str(series.dtype) != valid_dtype:
+    if not _check_data_type_equality(str(series.dtype), valid_dtype):
         raise ValueError(
             f"dtype mismatch between Series dtype {series.dtype}, and {schema.logical_type} dtype, {valid_dtype}"
         )